feat: copy and paste snippets

+++ src/commands/deleteSnippet.ts
@@ -1,19 +1,37 @@
 import * as vscode from "vscode";
+import * as nls from "vscode-nls";
 import { ISnippet } from "..";
 import { CodeSnippetsService } from "../CodeSnippetsService";
 import getSnippetTextDocument from "../core/getSnippetTextDocument";
 import refreshAllView from "../views/refreshAllView";
 
-export default async (snippet: ISnippet) => {
-  if (!snippet.uri) {
-    return;
-  }
-  if (!snippet.name) {
+const localize = nls.loadMessageBundle();
+
+export default async (snippets: ISnippet[]) => {
+  if (snippets.length === 0) {
     return;
   }
 
+  let message = localize(
+    "batchDeleteMsg",
+    "Do you want to delete selected {0} snippets?",
+    snippets.length
+  );
+
+  if (snippets.length === 1) {
+    if (!snippets[0].name) {
+      return;
+    }
+
+    message = localize(
+      "deleteMsg",
+      "Do you want to delete snippet {0}?",
+      snippets[0].name
+    );
+  }
+
   const answer = await vscode.window.showWarningMessage(
-    `Do you want to delete snippet ${snippet.name}?`,
+    message,
     {
       modal: true,
     },
@@ -24,13 +42,25 @@ export default async (snippet: ISnippet) => {
     return;
   }
 
-  const textDocument = await getSnippetTextDocument({
-    snippetsUri: snippet.uri,
-  });
+  for (const snippet of snippets) {
+    if (!snippet.uri) {
+      continue;
+    }
 
-  const codeSnippetsService = new CodeSnippetsService(textDocument);
+    if (!snippet.name) {
+      continue;
+    }
 
-  await codeSnippetsService.delete(snippet.name);
+    const textDocument = await getSnippetTextDocument({
+      snippetsUri: snippet.uri,
+    });
+
+    const codeSnippetsService = new CodeSnippetsService(textDocument);
+
+    await codeSnippetsService.delete(snippet.name);
+  }
 
   refreshAllView();
+
+  return true;
 };
